pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=1.0.0",
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.2.1",
    "speechrecognition>=3.14.5",
]
